Bug #1408691 bsp: camera: Converity warn
[root cause  ] value type and condition in if statement
[changes     ] cam_scaler.c/dcam_core.c/isp_3dnr.c/isp_slice.c
[side effects] NA
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums512_1h10
[test case]:camera function
[reviewers   ] yangyang.liu
[change_type ] feature_bugfix
[tag_product ] common

Change the type of value and the if statement to fix warn.

+++ core/isp2.6/src/cam_scaler.c
@@ -1117,7 +1117,7 @@ static void camscaler_rds_weight_normalize(
 	int64_t tmp_sum = 0;
 	int temp_value;
 	int64_t tmp_weight64 = 0;
-	u_int8_t i;
+	int i;
 	int sum = 0;
 	int diff;
